+++ src/apps/automator/src/helpers/replaceAutomatorTags.js
@@ -10,15 +10,11 @@ export default function replaceAutomatorTags(
   // Extra gap is necessary because some Markdown parsers are not handling
   // HTML comments well: https://github.com/jonschlinkert/remarkable/issues/209
   const gap = os.EOL + os.EOL;
-  const OPEN_TAG = `<!-- AUTOMATOR:${taskIdentifier} -->` + gap;
-  const CLOSE_TAG = gap + `<!-- /AUTOMATOR:${taskIdentifier} -->`;
+  const OPEN_TAG = `<!-- AUTOMATOR:${taskIdentifier} -->${gap}`;
+  const CLOSE_TAG = `${gap}<!-- /AUTOMATOR:${taskIdentifier} -->`;
 
   const TAGS_REGEXP = new RegExp(
-    '<!-- AUTOMATOR:' +
-      taskIdentifier +
-      ' -->[\\s\\S]*?<!-- /AUTOMATOR:' +
-      taskIdentifier +
-      ' -->',
+    `<!-- AUTOMATOR:${taskIdentifier} -->[\\s\\S]*?<!-- /AUTOMATOR:${taskIdentifier} -->`,
     'gu',
   );